requirements.txt
# pydantic<2.12
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
langchain-ollama
# chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
beautifulsoup4
langchain-classic
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
langchain_community